src/main/java/ru/job4j/concurrent/UserStorage.java
package ru.job4j.concurrent;


import net.jcip.annotations.GuardedBy;
import net.jcip.annotations.NotThreadSafe;

import java.util.HashMap;
import java.util.Map;

@NotThreadSafe
public class UserStorage {
    @GuardedBy("this")
    private final Map<Integer, User> users = new HashMap<>();

    public synchronized boolean add(User user) {
        return users.putIfAbsent(user.getId(), user) == null;
    }

    public synchronized boolean update(User user) {
        return users.replace(user.getId(), user) != null;
    }

    public synchronized boolean delete(User user) {
        return users.remove(user.getId(), user);
    }

    public synchronized boolean transfer(int fromId, int toId, int amount) {
        boolean rsl = false;
        User userFrom = users.get(fromId);
        User userTo = users.get(fromId);
        if ((userFrom != null) && (userTo != null) && (userFrom.getAmount() >= amount)) {
            userFrom.setAmount(userFrom.getAmount() - amount);
            userTo.setAmount(userTo.getAmount() + amount);
            rsl = true;
        }
        return rsl;
    }
}
